change upload materi

+++ src/components/pelajaran/materi.js
@@ -11,9 +11,18 @@ import "react-datepicker/dist/react-datepicker.css";
 
 import { Link } from 'react-router-dom';
 
-import { Modal, OverlayTrigger, Tooltip, Tabs, Tab, Button, Spinner } from 'react-bootstrap';
+import { Modal, OverlayTrigger, Tooltip, Tabs, Tab, Button, Spinner, DropdownButton, ButtonGroup } from 'react-bootstrap';
 import { MultiSelect } from 'react-sm-select';
 import 'react-sm-select/dist/styles.css';
+
+import '@trendmicro/react-dropdown/dist/react-dropdown.css';
+import Dropdown, {
+  MenuItem
+} from '@trendmicro/react-dropdown';
+
+// Be sure to include styles at some point, probably during your bootstraping
+import '@trendmicro/react-buttons/dist/react-buttons.css';
+import '@trendmicro/react-dropdown/dist/react-dropdown.css';
 
 import SocketContext from '../../socket';
 
@@ -194,44 +203,54 @@ class Overview extends React.Component {
     let cc = [...this.state.silabus];
     let item = cc.filter((item, i) => i === index);
 
-    let form = {
-      companyId: Storage.get('user').data.company_id,
-      pelajaranId: item[0].pelajaran_id,
-      number: item[0].sesi,
-      title: item[0].chapter_title,
-      content: item[0].chapter_body,
-      tatapmuka: item[0].tatapmuka,
-      tanggal: moment(item[0].start_date).format('YYYY-MM-DD HH:mm'),
-      silabusId: item[0].id
-    }
+    if(item[0].chapter_id) {
+      let form = {
+        companyId: Storage.get('user').data.company_id,
+        pelajaranId: item[0].pelajaran_id,
+        number: item[0].sesi,
+        title: item[0].chapter_title,
+        content: item[0].chapter_body,
+        tatapmuka: item[0].tatapmuka,
+        tanggal: moment(item[0].start_date).format('YYYY-MM-DD HH:mm'),
+        silabusId: item[0].id
+      }
 
-    API.post(`${API_SERVER}v2/pelajaran/chapter/create`, form).then(res => {
-      if (res.data.error) {
-        toast.warning(`Error: create chapter`)
-      }
-      else {
-        if (this.state.files) {
-          this.uplaodFiles(res.data.result.id, this.state.files)
-        } else {
-          toast.success(`Materi berhasil disimpan.`)
-          this.fetchOverview()
+      API.post(`${API_SERVER}v2/pelajaran/chapter/create`, form).then(res => {
+        if (res.data.error) {
+          toast.warning(`Error: create chapter`)
         }
-      }
-    })
+        else {
+          if (this.state.files) {
+            this.uplaodFiles(res.data.result.id, this.state.files)
+          } else {
+            toast.success(`Materi berhasil disimpan.`)
+            this.fetchOverview()
+          }
+        }
+      })
+    }
+    else {
+      toast.info(`Materi belum di upload.`)
+    }
   }
 
   deleteSesi(e, row, index) {
     let cc = [...this.state.silabus];
     let item = cc.filter((item, i) => i === index);
 
-    API.delete(`${API_SERVER}v2/pelajaran/chapter/delete/${item[0].chapter_id}`).then(res => {
-      if (res.data.error) {
-        toast.warning(`Error: delete chapter`)
-      }
-      else {
-        this.fetchOverview()
-      }
-    })
+    if(item[0].chapter_id) {
+      API.delete(`${API_SERVER}v2/pelajaran/chapter/delete/${item[0].chapter_id}`).then(res => {
+        if (res.data.error) {
+          toast.warning(`Error: delete chapter`)
+        }
+        else {
+          this.fetchOverview()
+        }
+      })
+    }
+    else {
+      toast.info(`Tidak ada yang dihapus.`)
+    }
   }
 
   render() {
@@ -256,13 +275,14 @@ class Overview extends React.Component {
                     <th style={{color: 'black'}} className="text-center">Periode</th>
                     <th style={{color: 'black'}} className="text-center">Durasi</th>
                     <th style={{color: 'black'}} className="text-center">Attachment</th>
+                    <th width="50px"></th>
                   </tr>
                 </thead>
                 <tbody>
                   {
                     this.state.loading &&
                     <tr>
-                      <td className="text-center" colSpan="7">fetching data...</td>
+                      <td className="text-center" colSpan="8">fetching data...</td>
                     </tr>
                   }
                   {
@@ -270,11 +290,9 @@ class Overview extends React.Component {
                       if(item.jenis === 0) {
                         return (
                           <>
-                            <tr key={i} style={{ cursor: 'pointer' }} data-toggle="collapse" data-target={`#collapse${i}`} data-parent="#myTableSilabus">
+                            <tr key={i} style={{ cursor: 'pointer' }}>
                               <td className="text-center">
-                                {item.sesi} <br/>
-                                <i onClick={e => this.copySesi(e, item, i)} style={{cursor: 'pointer'}} className="fa fa-copy mr-2" title="Copy"></i>
-                                <i onClick={e => this.deleteSesi(e, item, i)} style={{cursor: 'pointer'}} className="fa fa-trash mr-2" title="Delete"></i>
+                                {item.sesi}
                               </td>
                               <td>
                                 <OverlayTrigger
@@ -286,7 +304,7 @@ class Overview extends React.Component {
                                 </OverlayTrigger>
                               </td>
                               <td>{item.tujuan}</td>
-                              <td className="text-center">{item.start_date ? moment(item.start_date).format('DD/MM/YYYY HH:mm') : <span className="label label-primary">Upload Materi</span>}</td>
+                              <td className="text-center" data-toggle="collapse" data-target={`#collapse${i}`} data-parent="#myTableSilabus">{item.start_date ? moment(item.start_date).format('DD/MM/YYYY HH:mm') : <span className="label label-primary">Upload Materi</span>}</td>
                               <td  className="text-center">{item.periode}</td>
                               <td  className="text-center">{item.durasi} menit</td>
                               <td style={{padding: '12px'}} className="text-center">
@@ -294,9 +312,20 @@ class Overview extends React.Component {
                                   item.files ? <a href={item.files} target="_blank" className="silabus">Open</a> : '-'
                                 }
                               </td>
+                              <td>
+                                <Dropdown>
+                                  <Dropdown.Toggle btnStyle="flat" noCaret iconOnly>
+                                    <i className="fa fa-ellipsis-h"></i>
+                                  </Dropdown.Toggle>
+                                  <Dropdown.Menu>
+                                    <MenuItem onClick={e => this.copySesi(e, item, i)}><i className="fa fa-copy mr-2" title="Copy"></i> Copy</MenuItem>
+                                    <MenuItem onClick={e => this.deleteSesi(e, item, i)}><i className="fa fa-trash mr-2" title="Delete"></i> Delete</MenuItem>
+                                  </Dropdown.Menu>
+                                </Dropdown>
+                              </td>
                             </tr>
                             <tr className="collapse" id={`collapse${i}`}>
-                              <td colSpan="7">
+                              <td colSpan="8">
                                 <Tabs defaultActiveKey="materi" id="uncontrolled-tab-example">
                                   <Tab eventKey="materi" title="Materi" style={{padding: '8px'}}>
                                     <form className="row">
@@ -486,14 +515,12 @@ class Overview extends React.Component {
                       else {
                         return (
                           <>
-                            <tr key={i} style={{ cursor: 'pointer' }} data-toggle="collapse" data-target={`#collapse${i}`} data-parent="#myTableSilabus">
+                            <tr key={i} style={{ cursor: 'pointer' }}>
                               <td className="text-center">
-                                {item.sesi} <br/>
-                                <i onClick={e => this.copySesi(e, item, i)} style={{cursor: 'pointer'}} className="fa fa-copy" title="Copy"></i>
-                                <i onClick={e => this.deleteSesi(e, item, i)} style={{cursor: 'pointer'}} className="fa fa-trash mr-2" title="Delete"></i>
+                                {item.sesi}
                               </td>
                               <td colSpan="2" className="text-center">{item.jenis == 1 ? 'Kuis':'Ujian'}</td>
-                              <td className="text-center">{item.start_date ? moment(item.start_date).format('DD/MM/YYYY HH:mm') : <span className="label label-primary">Pilih {item.jenis == 1 ? 'Kuis':'Ujian'}</span>}</td>
+                              <td className="text-center" data-toggle="collapse" data-target={`#collapse${i}`} data-parent="#myTableSilabus">{item.start_date ? moment(item.start_date).format('DD/MM/YYYY HH:mm') : <span className="label label-primary">Pilih {item.jenis == 1 ? 'Kuis':'Ujian'}</span>}</td>
                               <td className="text-center">{item.periode}</td>
                               <td className="text-center">{item.durasi} menit</td>
                               <td className="text-center">
@@ -501,9 +528,20 @@ class Overview extends React.Component {
                                   item.files ? <a href={item.files} target="_blank" className="silabus">Open</a> : '-'
                                 }
                               </td>
+                              <td>
+                                <Dropdown>
+                                  <Dropdown.Toggle btnStyle="flat" noCaret iconOnly>
+                                    <i className="fa fa-ellipsis-h"></i>
+                                  </Dropdown.Toggle>
+                                  <Dropdown.Menu>
+                                    <MenuItem onClick={e => this.copySesi(e, item, i)}><i className="fa fa-copy mr-2" title="Copy"></i> Copy</MenuItem>
+                                    <MenuItem onClick={e => this.deleteSesi(e, item, i)}><i className="fa fa-trash mr-2" title="Delete"></i> Delete</MenuItem>
+                                  </Dropdown.Menu>
+                                </Dropdown>
+                              </td>
                             </tr>
                             <tr className="collapse" id={`collapse${i}`}>
-                              <td colSpan="7">
+                              <td colSpan="8">
 
                                 <form className="row">
                                   <div className="col-sm-6 bordered">
